Follow up fix on package renaming
