+++ .travis.yml
@@ -1,12 +1,9 @@
 language:
 - scala
 sudo: required
+dist: trusty
 jdk:
 - oraclejdk8
-addons:
-  apt:
-    packages:
-      - oracle-java8-installer
 scala:
 - 2.11.8
 cache:
