+++ conf/log4j.properties
@@ -80,7 +80,7 @@ log4j.threshhold=ALL
 #
 # Rolling File Appender
 #
-log4j.appender.RollingFileAppender=org.apache.gearpump.util.RecreateRollingFileAppender
+log4j.appender.RollingFileAppender=io.gearpump.util.RecreateRollingFileAppender
 log4j.appender.RollingFileAppender.File=${gearpump.log.dir}/${gearpump.log.file}
 log4j.appender.RollingFileAppender.checkFileInterval=60
 log4j.appender.RollingFileAppender.layout=org.apache.log4j.PatternLayout
@@ -101,7 +101,7 @@ log4j.appender.console.layout.ConversionPattern=[%c{1}] %m%n
 #
 # Application Log Appender
 #
-log4j.appender.ApplicationLogAppender=org.apache.gearpump.util.RecreateRollingFileAppender
+log4j.appender.ApplicationLogAppender=io.gearpump.util.RecreateRollingFileAppender
 log4j.appender.ApplicationLogAppender.File=${gearpump.application.log.dir}/${gearpump.application.log.file}
 log4j.appender.ApplicationLogAppender.checkFileInterval=60
