+++ core/src/main/resources/META-INF/services/io.gearpump.jarstore.JarStore
@@ -1,0 +1,1 @@
+io.gearpump.jarstore.local.LocalJarStore
