+++ examples/streaming/complexdag/src/test/scala/io/gearpump/streaming/examples/complexdag/SinkSpec.scala
@@ -17,30 +17,26 @@
  */
 package io.gearpump.streaming.examples.complexdag
 
-import akka.actor.ActorSystem
-import io.gearpump.cluster.TestUtil
+import io.gearpump.cluster.UserConfig
 import io.gearpump.streaming.MockUtil
-import io.gearpump.streaming.examples.complexdag.Source
+import org.scalatest.prop.PropertyChecks
+import org.scalatest.{BeforeAndAfter, Matchers, PropSpec}
 import io.gearpump.Message
-import io.gearpump.streaming.MockUtil._
-import org.mockito.Mockito._
-import org.scalatest.{Matchers, WordSpec}
 
-class SourceSpec extends WordSpec with Matchers {
+class SinkSpec extends PropSpec with PropertyChecks with Matchers with BeforeAndAfter {
 
-  "Source" should {
-    "Source should send a msg of Vector[String](classOf[Source].getCanonicalName)" in {
-      val system1 = ActorSystem("Source", TestUtil.DEFAULT_CONFIG)
+  val context = MockUtil.mockTaskContext
 
-      val system2 = ActorSystem("Reporter", TestUtil.DEFAULT_CONFIG)
+  val sink = new Sink(context, UserConfig.empty)
 
-      val context = MockUtil.mockTaskContext
+  property("Sink should send a Vector[String](classOf[Sink].getCanonicalName, " +
+    "classOf[Sink].getCanonicalName") {
+    val list = Vector(classOf[Sink].getCanonicalName)
+    val expected = Vector(classOf[Sink].getCanonicalName, classOf[Sink].getCanonicalName)
+    sink.onNext(Message(list))
 
-      val source = new Source(context, UserConfig.empty)
-      source.onNext(Message("start"))
-
-      verify(context).output(argMatch[Message](
-        Vector(classOf[Source].getCanonicalName) == _.value))
-    }
+    (0 until sink.list.size).map(i => {
+      assert(sink.list(i).equals(expected(i)))
+    })
   }
 }
