+++ examples/streaming/sol/src/main/scala/io/gearpump/streaming/examples/sol/SOL.scala
@@ -20,6 +20,7 @@ package io.gearpump.streaming.examples.sol
 
 import io.gearpump.cluster.UserConfig
 import io.gearpump.cluster.client.ClientContext
+import io.gearpump.cluster.main.{ArgumentsParser, CLIOption, ParseResult}
 import io.gearpump.streaming.partitioner.ShufflePartitioner
 import io.gearpump.util.{AkkaApp, Graph, LogUtil}
 import org.slf4j.Logger
