+++ examples/streaming/wordcount/src/main/scala/io/gearpump/streaming/examples/wordcount/dsl/WindowedWordCount.scala
@@ -24,6 +24,7 @@ import io.gearpump.streaming.dsl.scalaapi.{LoggerSink, StreamApp}
 import io.gearpump.streaming.source.{DataSource, Watermark}
 import io.gearpump.util.AkkaApp
 import io.gearpump.Message
+import io.gearpump.cluster.main.{ArgumentsParser, CLIOption}
 import io.gearpump.streaming.dsl.scalaapi.StreamApp
 import io.gearpump.streaming.dsl.window.api.{EventTimeTrigger, FixedWindows}
 import io.gearpump.streaming.source.Watermark
