+++ examples/streaming/wordcount/src/main/scala/io/gearpump/streaming/examples/wordcount/dsl/WordCount.scala
@@ -19,6 +19,7 @@
 package io.gearpump.streaming.examples.wordcount.dsl
 
 import io.gearpump.cluster.client.ClientContext
+import io.gearpump.cluster.main.{ArgumentsParser, CLIOption}
 import io.gearpump.streaming.dsl.scalaapi.StreamApp
 import io.gearpump.util.AkkaApp
 import io.gearpump.streaming.dsl.scalaapi.StreamApp._
