+++ services/dashboard/services/models/metrics_provider.js
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-angular.module('org.apache.gearpump.models')
+angular.module('io.gearpump.models')
 
   .service('MetricsProvider', function () {
     'use strict';
