+++ services/dashboard/services/restapi.js
@@ -233,7 +233,7 @@ angular.module('dashboard')
         replaceDagProcessor: function (files, formFormNames, appId, oldProcessorId, newProcessorDescription, inheritConf, onComplete) {
           var url = restapiV1Root + 'appmaster/' + appId + '/dynamicdag';
           var args = {
-            "$type": 'org.apache.gearpump.streaming.appmaster.DagManager.ReplaceProcessor',
+            "$type": 'io.gearpump.streaming.appmaster.DagManager.ReplaceProcessor',
             oldProcessorId: oldProcessorId,
             newProcessorDescription: angular.merge({
               id: oldProcessorId
